doc update to explain how to delete a node
Related to [this discussion about improving docs for deleting a node](https://github.com/RedisLabsModules/redis-graph/issues/158#issuecomment-433639549).

I think -- given the way the docs are structured...it seemed to me to put it under the `GRAPH.DELETE` section....primarily because that is the first place I looked at to find out how to delete a node -- then I noticed it was a whole graph delete.  So, most people are going to look there, and having a minor note like this will help folks, I think.