Added max memory protection support
The max memory protection can be set on using MEMORY_PROTECTION configuration option (off by default).
If not used, the memory protection does not change any code path.
The memory protection is set to (redis maxmemory)*2, if memory usage pass this number the memory
protection callback is called. default callback will crash the server, it is possible to give
another callback to handle the max memory violation differently.
Cron job is used to update the max memory (as it might change on runtime).