Fix leak when MAX_QUEUED_QUERIES is used.
When trying to add an async graph free job to the writers queue,
the queue might be full (because of MAX_QUEUED_QUERIES configuration).
This will cause the free job to be lost which will cause the graph
to leak.

Added a force option to ThreadPools_AddWorkWriter so that the free
jobs will not be lost.