Parameters list does not end with ; anymore
Parameters passed with a Cypher query were ended with a semicolon previously, see for instance https://github.com/RedisGraph/RedisGraph/issues/1078. Now this seems to return an error: 
```
127.0.0.1:6379> GRAPH.QUERY test "CYPHER name = \"opa\";  RETURN $name"
(error) Error: query with more than one statement is not supported.
```
This updates the documentation to reflect that. Or is this a code bug instead of a documentation being outdated bug, and the semicolon should still be accepted? I'm anxiously waiting for #1839, btw.