Enable running the tests by enabling debug command.
Previously there were failures while running the tests, usually in cases when
the `redis-server` was installed via homebrew and/or it was impossible to
configure it via the config file, and which (of course) wasn't built with the
debug command being enabled by default.

This enables the debug command when running the `redis-server`, so the tests will
pass now.