Fix a bug with overwriting a node representation string.
A bug appeared whenever a node would be printed out with its properties and
there was a prefix in the string. It was caused by not taking into account
the number of bytes having already been written to the string before printing
out the properties of a node.

This commit changes the behaviour so that the number of bytes is taken into
account and properly writes the properties map into a string.