Fix the race-conditions and introduce a flow test.
The race condition would occur when a "GRAPH.INFO" command is issued which
is supposed to over all the collections of all the threads without locking.
To resolve this, an inverse rw-lock is introduced. An inverse because we
don't need to lock it for exclusive writes and parallel reads, but for
parallel writes and exclusive reads, as the internal structure of the info
object bears the lock-free guarantees for writing by using collections
either with a synchronisation primitive or accessing only per-thread objects.

Introduces a simple test file which partially tests the already implemented
behaviour of the new command.