Improve the precision of the tests for the GRAPH.INFO GET STAT.
Rewrites the test so that it uses the "GRAPH.INFO QUERIES PREV" for
getting the previous query durations and compares those with the
ones in the statistics: they should be equal.

The previous method was only comparing with the total execution time
with some error margin and wasn't as precise as it is now.