Correct the thread pool code.
1. Fixes the use-after-free caused by the threads accessing the job
queues that have already been destroyed. Now the job queues are
destroyed after the threads have exited, guaranteeing there will be
no access to those job queues anymore, meaning those can be safely
destroyed.

2. Uses atomic integers instead of volatile keyword as the latter
isn't thread safe and doesn't involve any memory ordering changes
from the ordinary rules, while the declared variables are shared
among the threads and are used concurrently.

3. Removes the loop counting until 100 milliseconds are passed.
This code is useless when the pthread_join is used, as it will
wait anyway until the thread is "done", what, in turn, is already
ensured by waking up the threads and preventing them from entering
their next main loop iteration.

4. Uses pthread_join to join the threads before destroying them.
It is required to join attached threads before destroying those.
Not doing those will lead to a crash caused by a signal from
the kernel that will mean that there are still threads running while
the main thread (with TID==PID) has already finished. As the threads
should not be detached, the pthread_join must be used.

5. Removes the unnecessary pthread_detach use. The threads can't
be detached as they use a shared resource owned by the main thread,
and destroyed by it as well. The threads may only be detached when
they own all the resources they have themselves, yet even so it is
not a good practice nowadays and almost in all the cases the
attached thread with join is used anyway. Hence, the pthread_join
is also introduced to properly wait on the threads created.

6. Removes the excess uses of mutex and moves the variables to
atomic ones as those are wrote to and read from concurrently.
This should improve the performance as the atomic operations
are fully done within the same (user) address space, requires no
context switching and does not involve any system calls.