Use one queue for the working queries.
Using two separate queues was useful in the beginning, but in the
end, knowing that one thread can only execute one query at a time,
and so has this query being in one single mutually exclusive state,
having two separate collections is unnecessary.

This commit merges two queues for executing/reporting queries into
one - "working queries" queue.

This reduces the memory footprint for each graph but also introduces
a byte extra for each query to store as a state, so still much less
than the sizeof(QueryInfo), even though the magnitude of the queries
is small (which is equal to the number of threads).