Optimise the use of an attribute set.
Whenever the attribute set objects were used, we allocated excessive
memory, occupied larger registers, performed wider operations with
those when we didn't need to do that and avoid the extra costs of those.
At the same time, the postfix increments were used instead of the prefix
ones.

This commit brings those to justice and attempts to avoid the overhead,
providing the compiler with more precise information about the types
used and doing only what is required in the code.