Add a circular buffer without the read guarantee.
There are mainly two types of ring buffers: one with and one without the
read guarantee. The read guarantee means that the reader must be able to
read the data it hasn't read yet and this data can't be overwritten by
the write mechanism: in case the ring buffer has never been read and the
writes are exceeding the allocated capacity of this buffer, the writes
are refused. Excluding the read guarantee allows such writes to proceed
and overwrite the data the reader hasn't read yet by also adjusting the
read pointer to the least recently read entry if possible.

For the GRAPH.INFO we need to allocate a ring buffer of a pre-determined
capacity and we don't need to guarantee the reads. This allows us to
update the ring buffer as often and as much as possible without relying
on the reads.

For the ring buffer with the read guarantee there is already a
"circular_buffer.c", while this adds one without the read guarantee
under the name of "circular_buffer_nrg.c" where the "nrg" prefix means
"No Read Guarantee".