Fix the flaky cron abort later test.
The test should have been flaky due to the time limit imposed with no
control over the clock used as well as the scheduling.

This rewrites the test so that we don't rely on any clock rather than
the real clock and for just the time out, not for "letting something
happen". The test only proceeds when the cron thread has performed the
job and notifies the main thread about its completion. This is made
possible by using the file descriptors and the select system call.

The select also makes use of the time out facility and uses a three
second time interval for the test timeout.